Convert to UByte when reading single byte to fit contract

+++ app/src/main/kotlin/kttp/io/DefaultInputStream.kt
@@ -6,7 +6,7 @@ abstract class DefaultInputStream: InputStream() {
     override fun read(): Int {
         val bytes = ByteArray(1)
         val read = read(bytes)
-        return if (read == -1) -1 else bytes[0].toInt()
+        return if (read == -1) -1 else bytes[0].toUByte().toInt()
     }
 
     abstract override fun read(bytes: ByteArray, offset: Int, length: Int): Int
